Began programming and added statistical analysis section to text UI
21/7/24:
Began program by first importing the sample code and adjusting it to suit my program to see how everything works. Stumbled upon error when reading dataset (errno2) so for now switched it to read an absolute directory.

Data visualisation function was not plotting so attempted debugging it by printing something before and after trying to plot as well as adding an error message to understand what was wrong. Error message gave 'invalid numerical data'. Error ended up being with dataset as it contained fields in the first line, so after updating it would contain 2 rows of fields. Fixed by removing the fields within the dataset .csv file.

Also substituted the dataset with a copy that only contains 2 fields and 10 rows to ease testing and debugging process.

Added 'getAnalysis' function with the purpose of inputting whether the user would want to get the mean, median, mode, max or min.
Added 'getStatistical_Data' function with the purpose of calculating mean, median, mode, max and min.